🔄 refactor: Consolidate home page rendering and add health check endpoint
